# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is the Host API for Maestro System",
    "version" : "1.0.0",
    "title" : "MaestroHostService",
    "contact" : {
      "email" : "febo74@gmail.com"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/TheEdgeCompany/MaestroHostService/1.0.0",
...
  "schemes" : [ "https" ],
  "consumes" : [ "application/json", "application/octet-stream" ],
  "produces" : [ "application/json", "application/octet-stream" ],
  "security" : [ {
    "BasicAuth" : [ ]
  } ],
  "paths" : {
...
    "/Files/{fileId}" : {
      "get" : {
        "tags" : [ "Files" ],
        "description" : "Return a given file stream.",
        "operationId" : "GetFile",
        "parameters" : [ {
          "name" : "fileId",
          "in" : "path",
          "description" : "The id of the file to download.",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "offset",
          "in" : "query",
          "description" : "The optional starting point of a file segment. If this is absent or 0 the starting point is the beginning of the file.",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "length",
          "in" : "query",
          "description" : "The optional length of the file. If this is absent or 0 the entire file length is considered.",
          "required" : false,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/FileResponse"
            }
          },
          "400" : {
            "description" : "Bad request",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "404" : {
            "description" : "The file does not exists."
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Files" ],
        "description" : "Upload a given file stream.",
        "operationId" : "PostFile",
        "parameters" : [ {
          "name" : "fileId",
          "in" : "path",
          "description" : "The id of the file to download.",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "owner",
          "in" : "path",
          "description" : "The owner of the file.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "ownerId",
          "in" : "path",
          "description" : "The id of the owner.",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/FileResponse"
            }
          },
          "400" : {
            "description" : "Bad request",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "404" : {
            "description" : "The file does not exists."
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "securityDefinitions" : {
    "BasicAuth" : {
      "description" : "Uses client authentication, client ID and secret needed to be sent in the request.",
      "type" : "basic"
    }
  },
  "definitions" : {
...
    "FileResponse" : {
      "type" : "object",
      "properties" : {
        "data" : {
          "$ref" : "#/definitions/FileData"
        }
      },
      "description" : "The File response",
      "example" : {
        "data" : {
          "data" : {
            "size" : 1,
            "name" : "name",
            "sceneId" : 5,
            "contentId" : 5,
            "description" : "description",
            "type" : { },
            "content" : "content",
            "fileId" : 6
          }
        }
      }
    },
    "FileData" : {
      "type" : "object",
      "properties" : {
        "data" : {
          "$ref" : "#/definitions/File"
        }
      },
      "example" : {
        "data" : {
          "size" : 1,
          "name" : "name",
          "sceneId" : 5,
          "contentId" : 5,
          "description" : "description",
          "type" : { },
          "content" : "content",
          "fileId" : 6
        }
      }
    },
    "File" : {
      "type" : "object",
      "properties" : {
        "fileId" : {
          "type" : "integer",
          "description" : "The id of the file."
        },
        "name" : {
          "type" : "string",
          "description" : "The name of the file."
        },
        "description" : {
          "type" : "string"
        },
        "type" : {
          "$ref" : "#/definitions/ContentType"
        },
        "size" : {
          "type" : "integer",
          "description" : "The file size."
        },
        "sceneId" : {
          "type" : "integer",
          "description" : "The belonging scene."
        },
        "contentId" : {
          "type" : "integer",
          "description" : "The belonging content. If this is null, the file is the landmark and has only a belonging scene."
        },
        "content" : {
          "type" : "string",
          "format" : "base64",
          "description" : "The file content."
        }
      },
      "description" : "The file data.",
      "example" : {
        "size" : 1,
        "name" : "name",
        "sceneId" : 5,
        "contentId" : 5,
        "description" : "description",
        "type" : { },
        "content" : "content",
        "fileId" : 6
      }
    },
    "ContentType" : {
      "type" : "string",
      "description" : "The type of the file content.",
      "enum" : [ "text", "image", "video", "model" ],
      "x-enum-string" : true
    },
    "ErrorResponse" : {
      "type" : "object",
      "properties" : {
        "error" : {
          "$ref" : "#/definitions/Error"
        }
      },
      "description" : "Error response information."
    },
    "Error" : {
      "properties" : {
        "code" : {
          "type" : "integer",
          "description" : "Represents the code for thi error."
        },
        "message" : {
          "type" : "string",
          "description" : "A human readable message providing details about the error."
        },
        "errors" : {
          "type" : "array",
          "description" : "Container for any readable information.",
          "items" : {
            "$ref" : "#/definitions/Error_errors"
          }
        }
      },
      "description" : "Error information"
    },
    "Error_errors" : {
      "properties" : {
        "message" : {
          "type" : "string",
          "description" : "The message describing the error."
        }
      }
    }
  }
}